requirements.txt
Flask==3.0.1
Werkzeug==3.0.1
python-docx==1.1.0
Flask-WTF==1.2.1
WTForms==3.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
click==8.1.7
itsdangerous==2.1.2
lxml==5.1.0
python-dateutil==2.8.2
six==1.16.0
blinker==1.7.0